+++ CHANGELOG.md
@@ -1,1 +1,5 @@
 # Changelog
+
+## Unreleased
+
+- Show unsaved pages and blocks resolved from a Roam block reference in Manage search results.